requirements.txt
# Used as the OpenAI-compatible client for Mistral's chat completions endpoint.
openai>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0
rich>=13.0.0
pytest>=7.0.0
